+++ packages/main/test/specs/Slider.spec.js
@@ -216,269 +216,3 @@ describe("Slider elements - tooltip, step, tickmarks, labels", () => {
 		assert.strictEqual(await slider.getProperty("value"), 13, "current value should not be stepped to the next step (14)");
 	});
 });
-
-describe("Testing events", () => {
-
-	it("Should fire input event on use interaction and change event after user interaction finish", async () => {
-		const slider = await browser.$("#test-slider");
-		const eventResultSlider = await browser.$("#test-result-slider");
-
-		await slider.click();
-
-		assert.strictEqual(await eventResultSlider.getProperty("value") , 3, "Both input event and change event are fired after user interaction");
-	});
-
-	it("Should not fire change event after user interaction is finished if the current value is the same as the one at the start of the action", async () => {
-		const slider = await browser.$("#test-slider");
-		const eventResultSlider = await browser.$("#test-result-slider");
-
-		await slider.click();
-
-		assert.strictEqual(await eventResultSlider.getProperty("value") , 3, "Change event is not fired if the value is the same as before the start of the action");
-	});
-});
-
-describe("Accessibility", async () => {
-	it("Aria attributes are set correctly", async () => {
-		const slider = await browser.$("#basic-slider");
-		const sliderHandle = await slider.shadow$(".ui5-slider-handle");
-		const sliderId = await slider.getProperty("_id");
-
-		assert.strictEqual(await sliderHandle.getAttribute("aria-labelledby"),
-			`${sliderId}-accName ${sliderId}-sliderDesc`, "aria-labelledby is set correctly");
-		assert.strictEqual(await sliderHandle.getAttribute("aria-valuemin"),
-			`${await slider.getProperty("min")}`, "aria-valuemin is set correctly");
-		assert.strictEqual(await sliderHandle.getAttribute("aria-valuemax"),
-			`${await slider.getProperty("max")}`, "aria-valuemax is set correctly");
-		assert.strictEqual(await sliderHandle.getAttribute("aria-valuenow"),
-			`${await slider.getProperty("value")}`, "aria-valuenow is set correctly");
-	});
-
-	it("Click anywhere in the Slider should focus the Slider's handle", async () => {
-		await browser.url(`test/pages/Slider.html`);
-
-		const slider = await browser.$("#basic-slider");
-		const sliderHandle = await slider.shadow$(".ui5-slider-handle");
-
-		await slider.click();
-
-		const innerFocusedElement = await browser.custom$("activeElement", "#basic-slider");
-
-		assert.ok(await slider.isFocused(), "Slider component is focused");
-		assert.strictEqual(await browser.$(innerFocusedElement).getAttribute("class"), await sliderHandle.getAttribute("class"), "Slider handle has the shadowDom focus");
-	});
-
-	it("Tab should focus the Slider and move the visible focus outline to the slider's handle", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-		const sliderHandle = await slider.shadow$(".ui5-slider-handle");
-
-		await browser.keys("Tab");
-
-		const innerFocusedElement = await browser.custom$("activeElement", "#basic-slider-with-tooltip");
-
-		assert.ok(await slider.isFocused(), "Slider component is focused");
-		assert.strictEqual(await browser.$(innerFocusedElement).getAttribute("class"), await sliderHandle.getAttribute("class"), "Slider handle has the shadowDom focus");
-	});
-
-	it("Shift+Tab should focus the previous Slider and move the visible focus outline to the previous slider's handle", async () => {
-		const slider = await browser.$("#basic-slider");
-		const sliderHandle = await slider.shadow$(".ui5-slider-handle");
-
-		await browser.keys(["Shift", "Tab"]);
-
-		const innerFocusedElement = await browser.custom$("activeElement", "#basic-slider");
-
-		assert.ok(await slider.isFocused(), "Slider component is focused");
-		assert.strictEqual(await browser.$(innerFocusedElement).getAttribute("class"), await sliderHandle.getAttribute("class"), "Slider handle has the shadowDom focus");
-	});
-});
-
-
-describe("Accessibility: Testing keyboard handling", async () => {
-	it("Right arrow should increase the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider");
-
-		await slider.setProperty("value", 0);
-		await browser.keys("ArrowRight");
-
-		assert.strictEqual(await slider.getProperty("value"), 1, "Value is increased");
-	});
-
-	it("Left arrow should decrease the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider");
-
-		await browser.keys("ArrowLeft");
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is decreased");
-	});
-
-	it("Up arrow should increase the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider");
-
-		await browser.keys("ArrowUp");
-		assert.strictEqual(await slider.getProperty("value"), 1, "Value is increased");
-	});
-
-	it("Down arrow should increase the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider");
-
-		await browser.keys("ArrowDown");
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is decreased");
-	});
-
-	it("Ctrl + Right arrow should increase the value of the slider with a big increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys("Tab");
-		await browser.keys(["Control", "ArrowRight"]);
-
-		assert.strictEqual(await slider.getProperty("value"), 2, "Value is increased");
-	});
-
-	it("Ctrl + Left arrow should decrease the value of the slider with a big increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys(["Control", "ArrowLeft"]);
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is decreased");
-	});
-
-	it("Ctrl + Up arrow should increase the value of the slider with a big increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys(["Control", "ArrowUp"]);
-		assert.strictEqual(await slider.getProperty("value"), 2, "Value is increased");
-	});
-
-	it("Ctrl + Down arrow should increase the value of the slider with a big increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys(["Control", "ArrowDown"]);
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is decreased");
-	});
-
-	it("PageUp should increase the value of the slider with a big increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys("PageUp");
-		assert.strictEqual(await slider.getProperty("value"), 2, "Value is increased");
-	});
-
-	it("PageDown should decrease the value of the slider with a big increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys("PageDown");
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is decreased");
-	});
-
-	it("A '+' key press should increase the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys("+");
-		assert.strictEqual(await slider.getProperty("value"), 1, "Value is increased");
-	});
-
-	it("A '-' key press should decrease the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys("-");
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is decreased");
-	});
-
-	it("A numpad '+' key press should increase the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-		const numpadAdd = "\uE025";
-
-		await browser.keys(numpadAdd);
-		assert.strictEqual(await slider.getProperty("value"), 1, "Value is increased");
-	});
-
-	it("A numpad '-' key press should decrease the value of the slider with a small increment step", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-		const numpadSubtract = "\uE027";
-
-		await browser.keys(numpadSubtract);
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is decreased");
-	});
-
-	it("An 'End' key press should increase the value of the slider to its max", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys("End");
-		assert.strictEqual(await slider.getProperty("value"), 20, "Value is decreased");
-	});
-
-	it("A 'Home' key press should set the value of the slider to its minimum", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await browser.keys("Home");
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is increased");
-	});
-
-	it("A 'Esc' key press should return the value of the slider at its initial point at the time of its focusing", async () => {
-		const slider = await browser.$("#basic-slider-with-tooltip");
-
-		await slider.setProperty("value", 12);
-
-		await browser.keys("Escape");
-		assert.strictEqual(await slider.getProperty("value"), 0, "Value is increased");
-	});
-});
-
-describe("Testing resize handling and RTL support", () => {
-	it("Testing RTL support", async () => {
-		const slider = await browser.$("#basic-slider-rtl");
-		const sliderHandle = await slider.shadow$(".ui5-slider-handle");
-
-		assert.strictEqual((await sliderHandle.getCSSProperty("right")).value, "0px", "Initially if no value is set, the Slider handle is at the right of the Slider");
-
-		await slider.setProperty("value", 3);
-		assert.strictEqual(await sliderHandle.getAttribute("style"), "right: 30%;", "Slider handle should be 30% from the right");
-
-		await slider.click();
-
-		assert.strictEqual(await sliderHandle.getAttribute("style"), "right: 50%;", "Slider handle should be in the middle of the slider");
-		assert.strictEqual(await slider.getProperty("value"), 5, "Slider current value should be 5");
-
-		await sliderHandle.dragAndDrop({ x: -300, y: 1 });
-
-		assert.strictEqual(await sliderHandle.getAttribute("style"), "right: 80%;", "Slider handle should be 80% from the right of the slider");
-		assert.strictEqual(await slider.getProperty("value"), 8, "Slider current value should be 8");
-
-		await sliderHandle.dragAndDrop({ x: -100, y: 1 });
-
-		assert.strictEqual(await sliderHandle.getAttribute("style"), "right: 90%;", "Slider handle should be 90% from the right");
-		assert.strictEqual(await slider.getProperty("value"), 9, "Slider current value should be 9");
-
-		await sliderHandle.dragAndDrop({ x: -150, y: 1 });
-
-		assert.strictEqual(await sliderHandle.getAttribute("style"), "right: 100%;", "Slider handle should be at the left of the slider and not beyond its boundaries");
-		assert.strictEqual(await slider.getProperty("value"), 10, "Slider current value should be 10");
-	});
-
-	it("Testing RTL KBH support", async () => {
-		const slider = await browser.$("#basic-slider-rtl");
-		const sliderHandle = await slider.shadow$(".ui5-slider-handle");
-
-		await slider.setProperty("value", 0);
-		assert.strictEqual((await sliderHandle.getCSSProperty("right")).value, "0px", "Initially if no value is set, the Slider handle is at the right of the Slider");
-
-		await slider.keys("ArrowLeft");
-		await slider.keys("ArrowLeft");
-
-		assert.strictEqual(await sliderHandle.getAttribute("style"), "right: 20%;", "Slider handle should be 20% from the right of the slider");
-		assert.strictEqual(await slider.getProperty("value"), 2, "Slider current value should be 2");
-
-		await slider.keys("ArrowRight");
-
-		assert.strictEqual(await sliderHandle.getAttribute("style"), "right: 10%;", "Slider handle should be 10% from the right of the slider");
-		assert.strictEqual(await slider.getProperty("value"), 1, "Slider current value should be 1");
-	});
-
-	it("Should hide all labels except the first and the last one, if there is not enough space for all of them", async () => {
-		const slider = await browser.$("#slider-tickmarks-tooltips-labels");
-
-		await browser.setWindowSize(400, 2000);
-
-		assert.ok(await slider.getProperty("_labelsOverlapping"), "state should reflect if any of the labels is overlapping with another");
-		assert.ok(await slider.getProperty("_hiddenTickmarks"), "state should reflect if the tickmarks has less than 8px space between each of them");
-	});
-});
